ui: main: move home
