+++ src/App.js
@@ -10,9 +10,10 @@ import './comps/global/FontAwesome'
 import config from './data/config'
 import server from './data/server'
 import Navbar from './comps/global/Navbar'
-import Home from './comps/Home'
-import Login from './comps/Login'
+import Home from './comps/main/Home'
+import Login from './comps/main/Login'
 import Profile from './comps/account/Profile'
+import Shop from './comps/account/Shop'
 import Dashboard from './comps/account/Dashboard'
 import Footer from './comps/global/Footer'
 import Send from './comps/account/Send'
@@ -123,6 +124,7 @@ export default withRouter(class App extends Component {
           <Route exact path='/' component={Home} />
           <Route path='/login' render={routeProps => <Login {...routeProps} uiConfig={this.firebaseUIConfig} firebaseAuth={this.firebaseAuth}/>} />
           <Route path='/profile' render={routeProps => <Profile {...routeProps} user={this.state.user}/>} />
+          <Route path='/shop' component={Shop} />
           <Route path='/dashboard' render={routeProps => <Dashboard {...routeProps} user={this.state.user} acctQr={this.state.acctQr} userDoc={this.state.userDoc} mjMetaDoc={this.state.mjMetaDoc}/>} />
           <Route path='/send' render={routeProps => <Send {...routeProps} userDoc={this.state.userDoc}/>} />
           <Route path='/receive' render={routeProps => <Receive {...routeProps} user={this.state.user} acctQr={this.state.acctQr}/>} />
