--- src/comps/Tech.js
@@ -1,3 +1,0 @@
-// summary (media items)
-
-// details
