+++ src/comps/account/Shop.js
@@ -1,0 +1,8 @@
+import React from 'react'
+
+export default props =>
+  <React.Fragment>
+    <div className="mj-box">
+      <span>Wow</span>
+    </div>
+  </React.Fragment>
