+++ src/comps/main/Tech.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default props =>
+  <React.Fragment>
+    <div className="mj-box">
+      <span>Wow</span>
+    </div>
+  </React.Fragment>
+
+// summary (media items)
+
+// details (paragraphs)
